Use dedicated Subtitle Form component
